+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * **\[core\]** Rename the spy event from `ReactionStartedEvent` to `ObserveStartedEvent`.
 * **\[core\]** Rename the spy event from `ReactionCompletedEvent` to `ObserveCompletedEvent`.
 * **\[core\]** Rename the spy event from `ReactionObserveScheduled` to `ObserveScheduledEvent`.
+* **\[core\]** Enhance the `ComputeCompletedEvent` spy event to include the error message (if present) or
+  result computed.
 
 ### [v0.111](https://github.com/arez/arez/tree/v0.111) (2018-10-19)
 [Full Changelog](https://github.com/arez/arez/compare/v0.110...v0.111)